Documents/Rivaanah/rivaanah_bcakend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==1.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pymongo==3.11.4
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
"gunicorn" 
